+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Added `juliacall.TypeValue.__numpy_dtype__` attribute to allow converting Julia types
   to the corresponding NumPy dtype, like `numpy.dtype(jl.Int)`.
+* JuliaCall now launches Julia with 1 thread by default.
 * Bug fixes.
 
 ## 0.9.31 (2025-12-17)